Refactor common root elements
The Root component used by `ui.js` was nearly identical to
NotificationRoot, aside from the routes it used. The common parts have
been refactored into a separate Root component which is now used by
both.
