+++ app/scripts/ui.js
@@ -1,5 +1,5 @@
+import Root from '../../ui/app/pages/ui-root'
 const log = require('loglevel')
-const Root = require('../../ui/app/pages')
 const startPopup = require('./popup-core')
 
 startPopup(Root).catch(log.error)
